refactor: split the mobile terminal request and session files
Both files were one short edit from the 500-line limit, and both were
carrying two responsibilities.

`mobile_terminal_requests.rs` held the capability list a phone is told
about and the request allowlist it is admitted through, neither of which
is a terminal verb, and both of which grow with every mobile feature.
They move to `mobile_gateway_surface.rs`.

`terminal_session_controller.dart` held the session value type the whole
terminal feature passes around, separate from the controller that mints
it. It moves to `terminal_tab_session.dart`. Pure moves, no behaviour
change.
