+++ mobile/AGENTS.md
@@ -25,8 +25,8 @@ This file applies under `mobile/`. The root `AGENTS.md` also applies; this file 
 
 - The runtime gateway protocol version is `aleraMobileProtocolVersion` in `lib/src/core/mobile_protocol.dart`. The runtime enforces strict equality during `mobile.hello`, so never bump it for additive changes; feature-detect through the `runtimeCapabilities` array returned by the handshake instead.
 - Terminal output may arrive as a binary WebSocket message (`[u16be idLength][sessionId][raw bytes]`) instead of base64 inside JSON. The app asks for it with `binaryFrames: true` in `mobile.hello` and the runtime answers with what it granted; feature-detect through that response and `runtimeCapabilities`, never through `aleraMobileProtocolVersion`. The WebSocket already delimits messages, so there is no length-prefixed framing here, unlike the desktop socket.
-- Requests the gateway accepts from mobile clients are allowlisted in `rust/alera-cli/src/terminal_host/server/mobile_terminal_requests.rs` (`mobile_request_allowed`). Adding a new request type to the mobile app requires allowlisting it there first.
-- What `mobile.hello` advertises lives in `MOBILE_HELLO_CAPABILITIES` (`rust/alera-cli/src/terminal_host/server/mobile_terminal_requests.rs`), which is a different list from the one `status.get` returns. A capability the app feature-detects MUST be in the hello list: an omission is invisible, leaves every phone permanently on the older code path, and has no version to blame because the runtime requires an exact `aleraMobileProtocolVersion` match.
+- Requests the gateway accepts from mobile clients are allowlisted in `rust/alera-cli/src/terminal_host/server/mobile_gateway_surface.rs` (`mobile_request_allowed`). Adding a new request type to the mobile app requires allowlisting it there first.
+- What `mobile.hello` advertises lives in `MOBILE_HELLO_CAPABILITIES` (`rust/alera-cli/src/terminal_host/server/mobile_gateway_surface.rs`), which is a different list from the one `status.get` returns. A capability the app feature-detects MUST be in the hello list: an omission is invisible, leaves every phone permanently on the older code path, and has no version to blame because the runtime requires an exact `aleraMobileProtocolVersion` match.
 - Remote runtime restart is negotiated through `runtimeHostRestartV1`. Mobile sends a soft `host.restart` first and force-retries only after a second destructive confirmation when the host reports active work. The host MUST launch the replacement sidecar itself with its effective configuration before the phone relies on normal connection recovery; a phone cannot start a process on the paired machine.
 - Cloud account enrollment is delegated through an already-authenticated paired runtime. The app sends its stable cloud installation id as additive `mobile.hello.cloudDeviceId`; `mobile.cloudEnrollment.create` takes that id from the authenticated connection and returns a short-lived code that the app redeems immediately. Do not send a different device id in the enrollment request or ask the user to copy the internal code.
 - After changing a central push subscription, ask the paired runtime to run `mobile.cloudSubscriptions.refresh`. That request carries no mobile-supplied count; the runtime reads the authoritative value from Cloud so lifecycle retention cannot depend on untrusted client state.
