+++ mobile/lib/src/features/terminal/presentation/workspace_tabs_screen.dart
@@ -8,6 +8,7 @@ import 'package:alera_mobile/src/features/codex_chat/presentation/mobile_codex_c
 import 'package:alera_mobile/src/features/codex_chat/application/mobile_codex_composer_draft_store.dart';
 import 'package:alera_mobile/src/features/terminal/application/tabs_controller.dart';
 import 'package:alera_mobile/src/features/terminal/application/terminal_session_controller.dart';
+import 'package:alera_mobile/src/features/terminal/application/terminal_tab_session.dart';
 import 'package:alera_mobile/src/features/terminal/presentation/terminal_keys_settings_screen.dart';
 import 'package:alera_mobile/src/features/terminal/presentation/terminal_tab_view.dart';
 import 'package:alera_mobile/src/features/workbench/application/workbench_providers.dart';
